Major rewrite of the database.
All wptType objects are now a dbWaypoint.
A dbWaypoint can have a link to a dbGroundspeak.
A dbWaypoint can be a waypoint or a cache or a waymark.

Renamed dbCache to dbWaypoint.
Renamed dbCacheGroup to dbGroup.
Renamed dbContainer(Type|Size) to dbContainer..
Renamed dbCacheSymbol to dbSymbol

A log is placed by a dbName
A log is placed in dbState in dbCountry (although dbCountry and dbState have no relation)

Currently it compiles....