Fix bug where program crashes on import if correct authorization keys aren’t present
The program crashes when importing if not all the appropriate authorization keys are present — but they aren’t supplied as part of the GitHub fork, and there isn’t any documentation on what keys are needed.

This set of changes at least keeps the program from crashing, so that the import will work and a new developer can get something working.