Fix error in Help->Notices message
The Help->Notices message that tells the user to download notices incorrectly referenced the local menu button as being on the top left, but it’s on the top right.

Also fix grammar of message (“downloaded” rather than “download” in first sentence).

I changed the “key” for the message, as it serves both as the “key” and the “comment” of what the message means, and I thought it was cleaner to change the “key” so that if the translator assumed that the “key” was the correct message, then they would translate the message correctly.

I updated the “key” for the Dutch translation, but don’t know whether the dutch text needs to change as well.