https://github.com/MavEtJu/Geocube/pull/1
NightFlyer commented 4 days ago

There is a bug in Geocube where the local menu button isn't always hidden or shown when it is supposed to be. The problem is caused because viewWillAppear doesn't set menuGlobal.menuLocalButton to the appropriate localMenuButton, so sometimes the previous MHTabBarController's localMenuButton is being displayed (and so it doesn't get hidden).

You can see this bug if you switch between different tab bar controllers via the global menu and between the items within a tab bar controller which should show/hide the local menu button.

Here is one way to reproduce the problem:

    Go to navigate/target.
    Exit app (and kill) -- this just makes sure that you're in a known state when starting up.
    Start up app
    You'll be at navigate/compass (no local menu button)
    Choose target tab (local menu button appears)
    Use global menu to switch to help. You'll be on "About", with no local menu button
    Choose Help tab (local menu button appears)
    Use global menu to switch to navigate. You'll be on "Target" with a local menu button -- but I believe it's actually the local menu button from the Help tab controller, although visually you can't tell.
    Choose compass tab.
    Local menu is still there.

The fix is easy: just set the globalMenu object's cached value appropriately when the view is displayed.

My code does that (as well as printing out a log message for you to see what is happening).