Change display date format to better match date picker
The date picker uses words for month, and is in a localized order (month-day-year for US, day-month-year for other areas). The previous date display (2017-12-31) doesn’t match ever match the date picker (since it uses a month number).

I think it’s cleaner to use a more similar format for both the display and the date picker.