Have the definition of SecRandomCopyBytes match!
In one place, SecRandomCopyBytes was definited to take a void *, and in another place to take a uint8_t*. The compiler was complaining and so this code wouldn’t compile.

I changed the definition to match the Apple definition (although now if the Apple version isn’t available, it will still call it’s own version which has uint8_t* as the expected argument type, but I don’t think that is ia problem).