So far the great "let's modularize this" experiment.
Which slowed down xcode and caused a bouncy ball when you touched the mouse.
Which caused a bouncy ball when you tried to compile.
Which caused a bouncy ball when you to swap to the compiling tab.
Which caused a bouncy ball when you to swap back to the files tab.
Which caused a bouncy ball when you start to change text.

Maybe another time, but not today.